statev2: storage: tx: Add tx module and define high level tx
